Assignment_3/requirements.txt
google-generativeai>=0.3.2
langchain-community>=0.2.1
langchain-google-genai>=0.0.14      # ← NEW
faiss-cpu>=1.8.0
playwright>=1.43.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.2
datasets>=2.19.0
evaluate>=0.4.2
tiktoken>=0.6.0
